src/requirements.txt
alembic==1.14.1
amqp==5.3.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.3
Flask-Mail==0.10.0
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
Mako==1.3.9
MarkupSafe==2.1.5
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2